+++ forgerock-authenticator/CHANGELOG.md
@@ -1,6 +1,18 @@
-# Version 1.0.0
+# Version 1.1.0
+
+## [1.1.0]
+
+#### Updated
+
+- Android `forgerock-authenticator` dependency to the latest release.
+
+#### Added
+- New attributes on `PushNotification` model
+- New `PushType` model
+- New APIs to handle Biometric and Challenge push types
 
 ## [1.0.0]
 
 #### Added
 - Initial release of ForgeRock Authenticator for Flutter
+